solved problem 3 for small sizes

+++ Common/IProblem.cs
@@ -2,6 +2,6 @@
 {
     public interface IProblem
     {
-        int Solve();
+        long Solve();
     }
 }
